fix commands.
the node needs lt instead of relying on the controller node
username/apikey values.

Need to call getAllDevices() API in the controller as this is
rate limited. Push all the data down to the respective nodes.

Note that getAllDevices() is deprecated and we should be using
the webhooks api to get status updates.  But that's a bigger
change.